Version 2.0 Impliments random maze gen and overhauls UI
User can now generate mazes of any size they desire. Woah. Mazes created this way are randomly generated and are fully solveable/transposable, just like the premade ones. In randomly generated mazes, the starting position is currently set to always be the tile furthest away from the start.

UI has been recreated to be more "static" in look. Technically the console is still redrawing just as much, it just redraws ontop of previous assets now so it's much easier for the user to follow. UI also has the option to be in color (can be toggled off in settings).

Reworked transposition methods to properly function with new random mazes.

Lots of various tweaks/commenting/bug fixes.